+++ packages/compiler-pico-c/src/arch/index.ts
@@ -1,6 +1,7 @@
 import { CCompilerArch } from '../constants';
 import { CArchDescriptor } from './types';
 
+export * from './x86/asm-utils';
 import * as X86_16 from './x86/modes/16bit/sizeofPrimitiveType';
 
 const COMPILER_ARCH_DESCRIPTORS: Record<