+++ packages/compiler-pico-c/src/arch/x86/asm-utils/index.ts
@@ -3,3 +3,4 @@ export * from './genDefConst';
 export * from './genInstruction';
 export * from './genLabel';
 export * from './genMemAddress';
+export * from './wrapWithX86BootsectorAsm';