+++ packages/compiler-pico-c/src/arch/x86/index.ts
@@ -1,1 +1,2 @@
+export * from './asm-utils';
 export * from './backend/X86ArchBackend';