+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ts-c-compiler/cli
 
+## 1.0.1
+
+### Patch Changes
+
+- Add bootsector output option
+
 ## 1.0.0
 
 ### Major Changes