+++ packages/compiler-pico-c/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ts-c-compiler/compiler
 
+## 1.0.1
+
+### Patch Changes
+
+- Add bootsector output option
+- Updated dependencies
+  - @ts-c-compiler/core@1.0.1
+  - @ts-c-compiler/lexer@1.0.1
+  - @ts-c-compiler/rpn@1.0.1
+  - @ts-c-compiler/x86-assembler@1.0.1
+
 ## 1.0.0
 
 ### Major Changes